Working on simple connection to a SQLite database to store dashboards and telemetry data points

+++ src/simpleTimer.cpp
@@ -1,13 +1,14 @@
 #include "simpleTimer.hpp"
 #include <iostream>
 
-SimpleTimer::SimpleTimer(uint64_t interval) : m_interval(interval * 1000) {
+SimpleTimer::SimpleTimer(uint64_t interval)
+    : m_interval(interval * 1000) {
 }
 
 bool SimpleTimer::check() {
-    uint64_t currentTime = (system_clock::now().time_since_epoch()).count();
-    uint64_t delta = (currentTime - m_lastTime);
-    if (delta >= m_interval){
+    long long currentTime = (system_clock::now().time_since_epoch()).count();
+    long long delta       = (currentTime - m_lastTime);
+    if (delta >= m_interval) {
         m_lastTime = currentTime;
         return true;
     } else {
